third_party/WebKit/LayoutTests/fast/dom/custom/element-type.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<body>
<span is="x-x" id="a"></span>
<script>
description('Test that changing an instantiated custom element\'s \'is\' attribute does not affect its element type.');
shouldBeNonNull('window.testRunner', 'This test requires testRunner.');

// Some custom definition lookup happens at wrapper generation time so
// do the change in another world to avoid hiding the bug.
var worldId = 1;
testRunner.evaluateScriptInIsolatedWorld(
    worldId,
    'var a = document.querySelector("#a");' +
    'a.setAttribute("is", "x-y");');

var createdCallCount = 0;
var createdBy = '';
function onCreate(msg) {
  return function () {
    createdBy = msg;
    createdCallCount++;
  };
}

var protoX = Object.create(HTMLSpanElement.prototype);
protoX.createdCallback = onCreate('X');
var X = document.registerElement('x-x', {extends: 'span', prototype: protoX});

var protoY = Object.create(HTMLSpanElement.prototype);
protoY.createdCallback = onCreate('Y');
var Y = document.registerElement('x-y', {extends: 'span', prototype: protoY});

var a = document.querySelector('#a');
shouldBe('a.getAttribute("is")', '"x-y"');
shouldBeTrue('a instanceof X');
shouldBe('createdBy', '"X"');
shouldBe('createdCallCount', '1');
</script>
