third_party/WebKit/LayoutTests/fast/dom/custom/element-type-expected.txt

Test that changing an instantiated custom element's 'is' attribute does not affect its element type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.testRunner is non-null.
PASS a.getAttribute("is") is "x-y"
PASS a instanceof X is true
PASS createdBy is "X"
PASS createdCallCount is 1
PASS successfullyParsed is true

TEST COMPLETE
